+++ Sources/Views/MessagesCollectionView.swift
@@ -58,7 +58,7 @@ open class MessagesCollectionView: UICollectionView {
 
     public override init(frame: CGRect, collectionViewLayout layout: UICollectionViewLayout) {
         super.init(frame: frame, collectionViewLayout: layout)
-        backgroundColor = .backgroundColor
+        backgroundColor = .collectionViewBackground
         registerReusableViews()
         setupGestureRecognizers()
     }
@@ -80,6 +80,7 @@ open class MessagesCollectionView: UICollectionView {
         register(AudioMessageCell.self)
         register(ContactMessageCell.self)
         register(TypingIndicatorCell.self)
+        register(LinkPreviewMessageCell.self)
         register(MessageReusableView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader)
         register(MessageReusableView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionFooter)
     }