+++ Sources/Views/TypingIndicator.swift
@@ -33,7 +33,7 @@ open class TypingIndicator: UIView {
     public var bounceOffset: CGFloat = 2.5
     
     /// A convenience accessor for the `backgroundColor` of each dot
-    open var dotColor: UIColor = UIColor.darkTextColor {
+    open var dotColor: UIColor = UIColor.typingIndicatorDot {
         didSet {
             dots.forEach { $0.backgroundColor = dotColor }
         }